Minor refactorings

+++ src/main/java/me/robinvdb/runetracker/database/DataPointRepository.kt
@@ -3,4 +3,4 @@ package me.robinvdb.runetracker.database
 import me.robinvdb.runetracker.domain.DataPoint
 import org.springframework.data.repository.CrudRepository
 
-interface DataPointRepository : CrudRepository<DataPoint, Int>
+interface DataPointRepository : CrudRepository<DataPoint, Long>
