+++ src/main/java/me/robinvdb/runetracker/database/RunescapeUserRepository.kt
@@ -3,4 +3,4 @@ package me.robinvdb.runetracker.database
 import me.robinvdb.runetracker.domain.RunescapeUser
 import org.springframework.data.repository.CrudRepository
 
-interface RunescapeUserRepository : CrudRepository<RunescapeUser, Int>
+interface RunescapeUserRepository : CrudRepository<RunescapeUser, Long>
